+++ java/src/main/java/gherkin/formatter/FilterFormatter.java
@@ -148,7 +148,7 @@ public class FilterFormatter implements Formatter {
     }
 
     @Override
-    public void syntaxError(String state, String event, List<String> legalEvents, String uri, int line) {
+    public void syntaxError(String state, String event, List<String> legalEvents, String uri, Integer line) {
         throw new UnsupportedOperationException();
     }