+++ java/src/main/java/gherkin/formatter/JSONFormatter.java
@@ -167,7 +167,7 @@ public class JSONFormatter implements Reporter, Formatter {
     }
 
     @Override
-    public void syntaxError(String state, String event, List<String> legalEvents, String uri, int line) {
+    public void syntaxError(String state, String event, List<String> legalEvents, String uri, Integer line) {
         throw new UnsupportedOperationException();
     }