+++ java/src/main/java/gherkin/formatter/PrettyFormatter.java
@@ -374,7 +374,7 @@ public class PrettyFormatter implements Reporter, Formatter {
     }
 
     @Override
-    public void syntaxError(String state, String event, List<String> legalEvents, String uri, int line) {
+    public void syntaxError(String state, String event, List<String> legalEvents, String uri, Integer line) {
         throw new UnsupportedOperationException();
     }