+++ java/src/main/java/gherkin/formatter/model/DescribedStatement.java
@@ -5,7 +5,7 @@ import java.util.List;
 public abstract class DescribedStatement extends BasicStatement {
     private final String description;
 
-    public DescribedStatement(List<Comment> comments, String keyword, String name, String description, int line) {
+    public DescribedStatement(List<Comment> comments, String keyword, String name, String description, Integer line) {
         super(comments, keyword, name, line);
         this.description = description;
     }