+++ java/src/main/java/gherkin/formatter/model/Feature.java
@@ -5,7 +5,7 @@ import gherkin.formatter.Formatter;
 import java.util.List;
 
 public class Feature extends TagStatement {
-    public Feature(List<Comment> comments, List<Tag> tags, String keyword, String name, String description, int line, String id) {
+    public Feature(List<Comment> comments, List<Tag> tags, String keyword, String name, String description, Integer line, String id) {
         super(comments, tags, keyword, name, description, line, id);
     }