+++ java/src/main/java/gherkin/formatter/model/Scenario.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class Scenario extends TagStatement {
     private final String type = "scenario";
 
-    public Scenario(List<Comment> comments, List<Tag> tags, String keyword, String name, String description, int line, String id) {
+    public Scenario(List<Comment> comments, List<Tag> tags, String keyword, String name, String description, Integer line, String id) {
         super(comments, tags, keyword, name, description, line, id);
     }