+++ java/src/main/java/gherkin/lexer/Listener.java
@@ -3,25 +3,25 @@ package gherkin.lexer;
 import java.util.List;
 
 public interface Listener {
-    void comment(String comment, int line);
+    void comment(String comment, Integer line);
 
-    void tag(String tag, int line);
+    void tag(String tag, Integer line);
 
-    void feature(String keyword, String name, String description, int line);
+    void feature(String keyword, String name, String description, Integer line);
 
-    void background(String keyword, String name, String description, int line);
+    void background(String keyword, String name, String description, Integer line);
 
-    void scenario(String keyword, String name, String description, int line);
+    void scenario(String keyword, String name, String description, Integer line);
 
-    void scenarioOutline(String keyword, String name, String description, int line);
+    void scenarioOutline(String keyword, String name, String description, Integer line);
 
-    void examples(String keyword, String name, String description, int line);
+    void examples(String keyword, String name, String description, Integer line);
 
-    void step(String keyword, String name, int line);
+    void step(String keyword, String name, Integer line);
 
-    void row(List<String> cells, int line);
+    void row(List<String> cells, Integer line);
 
-    void docString(String contentType, String content, int line);
+    void docString(String contentType, String content, Integer line);
 
     void eof();
 }