+++ java/src/main/java/gherkin/parser/FormatterListener.java
@@ -73,63 +73,63 @@ public class FormatterListener implements Listener {
     }
 
     @Override
-    public void comment(String comment, int line) {
+    public void comment(String comment, Integer line) {
         stash.comment(new Comment(comment, line));
     }
 
     @Override
-    public void tag(String tag, int line) {
+    public void tag(String tag, Integer line) {
         stash.tag(new Tag(tag, line));
     }
 
     @Override
-    public void feature(String keyword, String name, String description, int line) {
+    public void feature(String keyword, String name, String description, Integer line) {
         formatter.feature(new Feature(stash.comments, stash.tags, keyword, name, description, line, stash.featureId(name)));
         stash.reset();
     }
 
     @Override
-    public void background(String keyword, String name, String description, int line) {
+    public void background(String keyword, String name, String description, Integer line) {
         formatter.background(new Background(stash.comments, keyword, name, description, line));
         stash.reset();
     }
 
     @Override
-    public void scenario(String keyword, String name, String description, int line) {
+    public void scenario(String keyword, String name, String description, Integer line) {
         replayStepsOrExamples();
         formatter.scenario(new Scenario(stash.comments, stash.tags, keyword, name, description, line, stash.featureElementId(name)));
         stash.reset();
     }
 
     @Override
-    public void scenarioOutline(String keyword, String name, String description, int line) {
+    public void scenarioOutline(String keyword, String name, String description, Integer line) {
         replayStepsOrExamples();
         formatter.scenarioOutline(new ScenarioOutline(stash.comments, stash.tags, keyword, name, description, line, stash.featureElementId(name)));
         stash.reset();
     }
 
     @Override
-    public void examples(String keyword, String name, String description, int line) {
+    public void examples(String keyword, String name, String description, Integer line) {
         replayStepsOrExamples();
         currentBuilder = new Examples.Builder(stash.comments, stash.tags, keyword, name, description, line, stash.examplesId(name));
         stash.reset();
     }
 
     @Override
-    public void step(String keyword, String name, int line) {
+    public void step(String keyword, String name, Integer line) {
         replayStepsOrExamples();
         currentBuilder = new Step.Builder(stash.comments, keyword, name, line);
         stash.reset();
     }
 
     @Override
-    public void row(List<String> cells, int line) {
+    public void row(List<String> cells, Integer line) {
         currentBuilder.row(stash.comments, cells, line, stash.nextExampleId());
         stash.reset();
     }
 
     @Override
-    public void docString(String contentType, String content, int line) {
+    public void docString(String contentType, String content, Integer line) {
         currentBuilder.docString(new DocString(contentType, content, line));
     }
 
@@ -142,7 +142,7 @@ public class FormatterListener implements Listener {
     /**
      * Not part of the API. Used for testing only.
      */
-    public void syntaxError(String state, String event, List<String> legalEvents, String uri, int line) {
+    public void syntaxError(String state, String event, List<String> legalEvents, String uri, Integer line) {
         formatter.syntaxError(state, event, legalEvents, uri, line);
     }