+++ java/src/main/java/gherkin/parser/StateMachineReader.java
@@ -23,37 +23,37 @@ public class StateMachineReader implements Listener {
         return transitionTable;
     }
 
-    public void tag(String tag, int line) {
+    public void tag(String tag, Integer line) {
     }
 
-    public void comment(String comment, int line) {
+    public void comment(String comment, Integer line) {
     }
 
-    public void feature(String keyword, String name, String description, int line) {
+    public void feature(String keyword, String name, String description, Integer line) {
     }
 
-    public void background(String keyword, String name, String description, int line) {
+    public void background(String keyword, String name, String description, Integer line) {
     }
 
-    public void scenario(String keyword, String name, String description, int line) {
+    public void scenario(String keyword, String name, String description, Integer line) {
     }
 
-    public void scenarioOutline(String keyword, String name, String description, int line) {
+    public void scenarioOutline(String keyword, String name, String description, Integer line) {
     }
 
-    public void examples(String keyword, String name, String description, int line) {
+    public void examples(String keyword, String name, String description, Integer line) {
     }
 
-    public void step(String keyword, String name, int line) {
+    public void step(String keyword, String name, Integer line) {
     }
 
-    public void docString(String contentType, String content, int line) {
+    public void docString(String contentType, String content, Integer line) {
     }
 
     public void eof() {
     }
 
-    public void row(List<String> cells, int line) {
+    public void row(List<String> cells, Integer line) {
         transitionTable.add(cells);
     }
 }